refactor: :bricks: Move game logic to BoardGame
Getting the API cards moves to the game page. The clickedStart state is initialized to true to indicate that the button should be displayed. Change the Dropdown activeClassName to activeclassname. Move the game logic that was in CardGame to BoardGame. The ul list that displays the cards is only shown if the level's cards are already ready.